add to Query and remove from Columns
